Fix selectize orientation filter feature tests
