+++ app/controllers/responsible/orientations_controller.rb
@@ -20,21 +20,19 @@ class Responsible::OrientationsController < Responsible::BaseController
     render :index
   end
 
-  def current_tcc_one(calendar = Calendar.current_by_tcc_one)
-    @title = orientation_calendar_title(calendar)
+  def current_tcc_one
+    @title = orientation_calendar_title(Calendar.current_by_tcc_one)
     add_breadcrumb @title, responsible_orientations_current_tcc_one_path
     @orientations = Orientation.by_current_tcc_one(params[:page], params[:term], params[:status])
     @search_url = responsible_orientations_search_current_tcc_one_path
-
     render :current_index
   end
 
-  def current_tcc_two(calendar = Calendar.current_by_tcc_two)
-    @title = orientation_calendar_title(calendar)
+  def current_tcc_two
+    @title = orientation_calendar_title(Calendar.current_by_tcc_two)
     add_breadcrumb @title, responsible_orientations_current_tcc_two_path
     @orientations = Orientation.by_current_tcc_two(params[:page], params[:term], params[:status])
     @search_url = responsible_orientations_search_current_tcc_two_path
-
     render :current_index
   end
